requirements.txt
Django==1.9.12
djangorestframework==3.6.3

